Fixed coverity warning
Warnings fixed: SIZEOF_MISMATCH

+++ nci/jni/HciEventManager.cpp
@@ -168,8 +168,7 @@ std::vector<uint8_t> HciEventManager::getDataFromBerTlv(
       return std::vector<uint8_t>(berTlv.begin() + 4, berTlv.end());
     }
   } else if (lengthTag == 0x84 && berTlv.size() > 5) {
-    size_t length =
-        (berTlv[1] << 24) | (berTlv[2] << 16) | (berTlv[3] << 8) | berTlv[4];
+    size_t length = (size_t)((berTlv[1] << 24) | (berTlv[2] << 16) | (berTlv[3] << 8) | berTlv[4]);
     if ((length + 5) == berTlv.size()) {
       return std::vector<uint8_t>(berTlv.begin() + 5, berTlv.end());
     }
